Factoring out the load behavior - Evaluate
