+++ SpiceSharp/Components/Semiconductors/MOS/Common/LoadBehavior.cs
@@ -16,6 +16,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Common
         private BaseConfiguration _baseConfig;
         private ModelBaseParameters _mbp;
         private BaseParameters _bp;
+        private TemperatureBehavior _temp;
 
         /// <summary>
         /// Gets or sets the drain current.
@@ -88,6 +89,24 @@ namespace SpiceSharp.Components.MosfetBehaviors.Common
         /// </value>
         [ParameterName("gbd"), ParameterInfo("Bulk-Drain conductance")]
         public double CondBd { get; protected set; }
+
+        /// <summary>
+        /// Gets or sets the turn-on voltage.
+        /// </summary>
+        /// <value>
+        /// The turn-on voltage.
+        /// </value>
+        [ParameterName("von"), ParameterInfo("Turn-on voltage")]
+        public double Von { get; protected set; }
+
+        /// <summary>
+        /// Gets or sets the saturation voltage.
+        /// </summary>
+        /// <value>
+        /// The saturation voltage.
+        /// </value>
+        [ParameterName("vdsat"), ParameterInfo("Saturation drain voltage")]
+        public double SaturationVoltageDs { get; protected set; }
 
         /// <summary>
         /// Extra variables
@@ -175,6 +194,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Common
             _baseConfig = simulation.Configurations.Get<BaseConfiguration>();
             _bp = provider.GetParameterSet<BaseParameters>();
             _mbp = provider.GetParameterSet<ModelBaseParameters>("model");
+            _temp = provider.GetBehavior<TemperatureBehavior>();
         }
 
         /// <summary>
@@ -265,16 +285,226 @@ namespace SpiceSharp.Components.MosfetBehaviors.Common
             SourcePrimeDrainPrimePtr = null;
         }
 
+        /// <summary>
+        /// Loads the Y-matrix and Rhs-vector.
+        /// </summary>
+        /// <param name="simulation">The base simulation.</param>
         public override void Load(BaseSimulation simulation)
         {
-            
+            if (simulation == null)
+                throw new ArgumentNullException(nameof(simulation));
+            var state = simulation.RealState;
+
+            double vgs, vds, vbs, vbd, vgd, von;
+            double drainSatCur, sourceSatCur;
+            int check = 1;
+
+            if (_temp.TempSaturationCurrentDensity.Equals(0) || _bp.DrainArea.Value <= 0 || _bp.SourceArea.Value <= 0)
+            {
+                drainSatCur = _temp.TempSaturationCurrent;
+                sourceSatCur = _temp.TempSaturationCurrent;
+            }
+            else
+            {
+                drainSatCur = _temp.TempSaturationCurrentDensity * _bp.DrainArea;
+                sourceSatCur = _temp.TempSaturationCurrentDensity * _bp.SourceArea;
+            }
+
+            var vt = Circuit.KOverQ * _bp.Temperature;
+
+            if (state.Init == InitializationModes.Float || (simulation is TimeSimulation tsim && tsim.Method.BaseTime.Equals(0.0)) ||
+                state.Init == InitializationModes.Fix && !_bp.Off)
+            {
+                // General iteration
+                vbs = _mbp.MosfetType * (state.Solution[BulkNode] - state.Solution[SourceNodePrime]);
+                vgs = _mbp.MosfetType * (state.Solution[GateNode] - state.Solution[SourceNodePrime]);
+                vds = _mbp.MosfetType * (state.Solution[DrainNodePrime] - state.Solution[SourceNodePrime]);
+
+                // now some common crunching for some more useful quantities
+                vbd = vbs - vds;
+                vgd = vgs - vds;
+                var vgdo = VoltageGs - VoltageDs;
+                von = _mbp.MosfetType * Von;
+
+                /*
+				 * limiting
+				 * we want to keep device voltages from changing
+				 * so fast that the exponentials churn out overflows
+				 * and similar rudeness
+				 */
+                // NOTE: Spice 3f5 does not write out Vgs during DC analysis, so DEVfetlim may give different results in Spice 3f5
+                if (VoltageDs >= 0)
+                {
+                    vgs = Transistor.LimitFet(vgs, VoltageGs, von);
+                    vds = vgs - vgd;
+                    vds = Transistor.LimitVoltageDs(vds, VoltageDs);
+                }
+                else
+                {
+                    vgd = Transistor.LimitFet(vgd, vgdo, von);
+                    vds = vgs - vgd;
+                    vds = -Transistor.LimitVoltageDs(-vds, -VoltageDs);
+                    vgs = vgd + vds;
+                }
+                if (vds >= 0)
+                {
+                    vbs = Transistor.LimitJunction(vbs, VoltageBs, vt, _temp.SourceVCritical, out check);
+                }
+                else
+                {
+                    vbd = Transistor.LimitJunction(vbd, VoltageBd, vt, _temp.DrainVCritical, out check);
+                    vbs = vbd + vds;
+                }
+            }
+            else
+            {
+                /* ok - not one of the simple cases, so we have to
+				 * look at all of the possibilities for why we were
+				 * called.  We still just initialize the three voltages
+				 */
+                if (state.Init == InitializationModes.Junction && !_bp.Off)
+                {
+                    vds = _mbp.MosfetType * _bp.InitialVoltageDs;
+                    vgs = _mbp.MosfetType * _bp.InitialVoltageGs;
+                    vbs = _mbp.MosfetType * _bp.InitialVoltageBs;
+
+                    // TODO: At some point, check what this is supposed to do
+                    if (vds.Equals(0) && vgs.Equals(0) && vbs.Equals(0) && (state.UseDc || !state.UseIc))
+                    {
+                        vbs = -1;
+                        vgs = _mbp.MosfetType * _temp.TempVt0;
+                        vds = 0;
+                    }
+                }
+                else
+                {
+                    vbs = vgs = vds = 0;
+                }
+            }
+
+            /*
+             * now all the preliminaries are over - we can start doing the
+             * real work
+             */
+            vbd = vbs - vds;
+            vgd = vgs - vds;
+
+            /*
+			 * bulk - source and bulk - drain diodes
+			 * here we just evaluate the ideal diode current and the
+			 * corresponding derivative (conductance).
+			 */
+            if (vbs <= 0)
+            {
+                CondBs = sourceSatCur / vt;
+                BsCurrent = CondBs * vbs;
+                CondBs += _baseConfig.Gmin;
+            }
+            else
+            {
+                var evbs = Math.Exp(Math.Min(Transistor.MaximumExponentArgument, vbs / vt));
+                CondBs = sourceSatCur * evbs / vt + _baseConfig.Gmin;
+                BsCurrent = sourceSatCur * (evbs - 1);
+            }
+            if (vbd <= 0)
+            {
+                CondBd = drainSatCur / vt;
+                BdCurrent = CondBd * vbd;
+                CondBd += _baseConfig.Gmin;
+            }
+            else
+            {
+                var evbd = Math.Exp(Math.Min(Transistor.MaximumExponentArgument, vbd / vt));
+                CondBd = drainSatCur * evbd / vt + _baseConfig.Gmin;
+                BdCurrent = drainSatCur * (evbd - 1);
+            }
+
+            /*
+             * Now to determine whether the user was able to correctly
+			 * identify the source and drain of his device
+			 */
+            if (vds >= 0)
+            {
+                // normal mode
+                Mode = 1;
+            }
+            else
+            {
+                // inverse mode
+                Mode = -1;
+            }
+
+            // Update
+            VoltageBs = vbs;
+            VoltageBd = vbd;
+            VoltageGs = vgs;
+            VoltageDs = vds;
+
+            // Evaluate the currents and derivatives
+            var cdrain = Mode > 0 ? Evaluate(vgs, vds, vbs) : Evaluate(vgd, -vds, vbd);
+            DrainCurrent = Mode * cdrain - BdCurrent;
+
+            // Check convergence
+            if (!_bp.Off || state.Init != InitializationModes.Fix)
+            {
+                if (check == 1)
+                    state.IsConvergent = false;
+            }
+
+            // Load current vector
+            double xnrm, xrev, cdreq;
+            var ceqbs = _mbp.MosfetType * (BsCurrent - (CondBs - _baseConfig.Gmin) * vbs);
+            var ceqbd = _mbp.MosfetType * (BdCurrent - (CondBd - _baseConfig.Gmin) * vbd);
+            if (Mode >= 0)
+            {
+                xnrm = 1;
+                xrev = 0;
+                cdreq = _mbp.MosfetType * (cdrain - CondDs * vds - Transconductance * vgs - TransconductanceBs * vbs);
+            }
+            else
+            {
+                xnrm = 0;
+                xrev = 1;
+                cdreq = -_mbp.MosfetType * (cdrain - CondDs * -vds - Transconductance * vgd - TransconductanceBs * vbd);
+            }
+            BulkPtr.Value -= ceqbs + ceqbd;
+            DrainPrimePtr.Value += ceqbd - cdreq;
+            SourcePrimePtr.Value += cdreq + ceqbs;
+
+            // Load Y-matrix
+            DrainDrainPtr.Value += _temp.DrainConductance;
+            SourceSourcePtr.Value += _temp.SourceConductance;
+            BulkBulkPtr.Value += CondBd + CondBs;
+            DrainPrimeDrainPrimePtr.Value += _temp.DrainConductance + CondDs + CondBd + xrev * (Transconductance + TransconductanceBs);
+            SourcePrimeSourcePrimePtr.Value += _temp.SourceConductance + CondDs + CondBs + xnrm * (Transconductance + TransconductanceBs);
+            DrainDrainPrimePtr.Value += -_temp.DrainConductance;
+            SourceSourcePrimePtr.Value += -_temp.SourceConductance;
+            BulkDrainPrimePtr.Value -= CondBd;
+            BulkSourcePrimePtr.Value -= CondBs;
+            DrainPrimeDrainPtr.Value += -_temp.DrainConductance;
+            DrainPrimeGatePtr.Value += (xnrm - xrev) * Transconductance;
+            DrainPrimeBulkPtr.Value += -CondBd + (xnrm - xrev) * TransconductanceBs;
+            DrainPrimeSourcePrimePtr.Value += -CondDs - xnrm * (Transconductance + TransconductanceBs);
+            SourcePrimeGatePtr.Value += -(xnrm - xrev) * Transconductance;
+            SourcePrimeSourcePtr.Value += -_temp.SourceConductance;
+            SourcePrimeBulkPtr.Value += -CondBs - (xnrm - xrev) * TransconductanceBs;
+            SourcePrimeDrainPrimePtr.Value += -CondDs - xrev * (Transconductance + TransconductanceBs);
         }
 
         /// <summary>
-        /// Check convergence
+        /// This routine evaluates the drain current and its derivatives. It also returns the calculated
+        /// drain current.
         /// </summary>
-        /// <param name="simulation">Base simulation</param>
-        /// <returns></returns>
+        protected abstract double Evaluate(double vgs, double vds, double vbs);
+
+        /// <summary>
+        /// Tests convergence at the device-level.
+        /// </summary>
+        /// <param name="simulation">The base simulation.</param>
+        /// <returns>
+        /// <c>true</c> if the device determines the solution converges; otherwise, <c>false</c>.
+        /// </returns>
+        /// <exception cref="ArgumentNullException">simulation</exception>
         public override bool IsConvergent(BaseSimulation simulation)
         {
 			if (simulation == null)
