+++ SpiceSharp/Components/Semiconductors/MOS/Level1/TemperatureBehavior.cs
@@ -8,7 +8,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level1
     /// <summary>
     /// Temperature behavior for a <see cref="Mosfet1"/>
     /// </summary>
-    public class TemperatureBehavior : BaseTemperatureBehavior
+    public class TemperatureBehavior : Common.TemperatureBehavior
     {
         /// <summary>
         /// Necessary behaviors
@@ -20,14 +20,6 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level1
         /// <summary>
         /// Shared variables
         /// </summary>
-        [ParameterName("sourcevcrit"), ParameterInfo("Critical source voltage")]
-        public double SourceVCritical { get; protected set; }
-        [ParameterName("drainvcrit"), ParameterInfo("Critical drain voltage")]
-        public double DrainVCritical { get; protected set; }
-        [ParameterName("sourceconductance"), ParameterInfo("Conductance of source")]
-        public double SourceConductance { get; protected set; }
-        [ParameterName("drainconductance"), ParameterInfo("Conductance of drain")]
-        public double DrainConductance { get; protected set; }
         [ParameterName("cbd0"), ParameterInfo("Zero-Bias B-D junction capacitance")]
         public double CapBd { get; protected set; }
         [ParameterName("cbdsw0"), ParameterInfo(" ")]
@@ -37,37 +29,13 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level1
         [ParameterName("cbssw0"), ParameterInfo(" ")]
         public double CapBsSidewall { get; protected set; }
 
-        [ParameterName("rs"), ParameterInfo("Source resistance")]
-        public double SourceResistance
-        {
-            get
-            {
-                if (SourceConductance > 0.0)
-                    return 1.0 / SourceConductance;
-                return 0.0;
-            }
-        }
-        [ParameterName("rd"), ParameterInfo("Drain conductance")]
-        public double DrainResistance
-        {
-            get
-            {
-                if (DrainConductance > 0.0)
-                    return 1.0 / DrainConductance;
-                return 0.0;
-            }
-        }
-
         /// <summary>
         /// Extra variables
         /// </summary>
-        public double TempTransconductance { get; protected set; }
+        
         public double TempSurfMob { get; protected set; }
         public double TempPhi { get; protected set; }
         public double TempVoltageBi { get; protected set; }
-        public double TempVt0 { get; protected set; }
-        public double TempSaturationCurrent { get; protected set; }
-        public double TempSaturationCurrentDensity { get; protected set; }
         public double TempCapBd { get; protected set; }
         public double TempCapBs { get; protected set; }
         public double TempJunctionCap { get; protected set; }
