+++ SpiceSharp/Components/Semiconductors/MOS/Level2/LoadBehavior.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using SpiceSharp.Attributes;
 using SpiceSharp.Behaviors;
 using SpiceSharp.Simulations;
 
@@ -17,21 +16,12 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level2
         private ModelBaseParameters _mbp;
         private TemperatureBehavior _temp;
         private ModelTemperatureBehavior _modeltemp;
-        private BaseConfiguration _baseConfig;
 
         /// <summary>
         /// Some signs used in the model
         /// </summary>
         private static readonly double[] Sig1 = { 1.0, -1.0, 1.0, -1.0 };
         private static readonly double[] Sig2 = { 1.0, 1.0, -1.0, -1.0 };
-
-        /// <summary>
-        /// Shared parameters
-        /// </summary>
-        [ParameterName("von"), ParameterInfo("Turn-on voltage")]
-        public double Von { get; protected set; }
-        [ParameterName("vdsat"), ParameterInfo("Saturation drain voltage")]
-        public double SaturationVoltageDs { get; protected set; }
 
         /// <summary>
         /// Constructor
@@ -49,9 +39,6 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level2
             if (provider == null)
                 throw new ArgumentNullException(nameof(provider));
             base.Setup(simulation, provider);
-
-            // Get configurations
-            _baseConfig = simulation.Configurations.Get<BaseConfiguration>();
 
             // Get parameters
             _bp = provider.GetParameterSet<BaseParameters>();
@@ -73,7 +60,6 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level2
         /// <param name="simulation"></param>
         public override void Unsetup(Simulation simulation)
         {
-            _baseConfig = null;
             _bp = null;
             _mbp = null;
             _temp = null;
@@ -85,157 +71,16 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level2
         /// <summary>
         /// Execute behavior
         /// </summary>
-        /// <param name="simulation">Base simulation</param>
-        public override void Load(BaseSimulation simulation)
+        protected override double Evaluate(double vgs, double vds, double vbs)
         {
-            if (simulation == null)
-                throw new ArgumentNullException(nameof(simulation));
-
-            var state = simulation.RealState;
-            double drainSatCur, sourceSatCur,
-                vgs, vds, vbs, vbd, vgd;
             double von;
-            double vdsat, cdrain = 0.0, cdreq;
-            int xnrm, xrev;
+            double vdsat, cdrain = 0.0;
 
             var vt = Circuit.KOverQ * _bp.Temperature;
-            var check = 1;
-
             var effectiveLength = _bp.Length - 2 * _mbp.LateralDiffusion;
-            if (_temp.TempSaturationCurrentDensity.Equals(0) || _bp.DrainArea.Value <= 0 || _bp.SourceArea.Value <= 0)
-            {
-                drainSatCur = _temp.TempSaturationCurrent;
-                sourceSatCur = _temp.TempSaturationCurrent;
-            }
-            else
-            {
-                drainSatCur = _temp.TempSaturationCurrentDensity * _bp.DrainArea;
-                sourceSatCur = _temp.TempSaturationCurrentDensity * _bp.SourceArea;
-            }
-
             var beta = _temp.TempTransconductance * _bp.Width / effectiveLength;
             var oxideCap = _mbp.OxideCapFactor * effectiveLength * _bp.Width;
-
-            if (state.Init == InitializationModes.Float || (simulation is TimeSimulation tsim && tsim.Method.BaseTime.Equals(0.0)) ||
-                state.Init == InitializationModes.Fix && !_bp.Off)
-            {
-                // General iteration
-                vbs = _mbp.MosfetType * (state.Solution[BulkNode] - state.Solution[SourceNodePrime]);
-                vgs = _mbp.MosfetType * (state.Solution[GateNode] - state.Solution[SourceNodePrime]);
-                vds = _mbp.MosfetType * (state.Solution[DrainNodePrime] - state.Solution[SourceNodePrime]);
-
-                // now some common crunching for some more useful quantities
-                vbd = vbs - vds;
-                vgd = vgs - vds;
-                var vgdo = VoltageGs - VoltageDs;
-                von = _mbp.MosfetType * Von;
-
-                /*
-				 * limiting
-				 * we want to keep device voltages from changing
-				 * so fast that the exponentials churn out overflows
-				 * and similar rudeness
-				 */
-                if (VoltageDs >= 0)
-                {
-                    vgs = Transistor.LimitFet(vgs, VoltageGs, von);
-                    vds = vgs - vgd;
-                    vds = Transistor.LimitVoltageDs(vds, VoltageDs);
-                }
-                else
-                {
-                    vgd = Transistor.LimitFet(vgd, vgdo, von);
-                    vds = vgs - vgd;
-                    vds = -Transistor.LimitVoltageDs(-vds, -VoltageDs);
-                    vgs = vgd + vds;
-                }
-                if (vds >= 0)
-                {
-                    vbs = Transistor.LimitJunction(vbs, VoltageBs, vt, _temp.SourceVCritical, out check);
-                }
-                else
-                {
-                    vbd = Transistor.LimitJunction(vbd, VoltageBd, vt, _temp.DrainVCritical, out check);
-                    vbs = vbd + vds;
-                }
-            }
-            else
-            {
-                /* ok - not one of the simple cases, so we have to
-				 * look at all of the possibilities for why we were
-				 * called.  We still just initialize the three voltages
-				 */
-                if (state.Init == InitializationModes.Junction && !_bp.Off)
-                {
-                    vds = _mbp.MosfetType * _bp.InitialVoltageDs;
-                    vgs = _mbp.MosfetType * _bp.InitialVoltageGs;
-                    vbs = _mbp.MosfetType * _bp.InitialVoltageBs;
-
-                    // TODO: At some point, check what this is supposed to do
-                    if (vds.Equals(0) && vgs.Equals(0) && vbs.Equals(0) && (state.UseDc || !state.UseIc))
-                    {
-                        vbs = -1;
-                        vgs = _mbp.MosfetType * _temp.TempVt0;
-                        vds = 0;
-                    }
-                }
-                else
-                {
-                    vbs = vgs = vds = 0;
-                }
-            }
-
-            /*
-			 * now all the preliminaries are over - we can start doing the
-			 * real work
-			 */
-            vbd = vbs - vds;
-            vgd = vgs - vds;
-
-            /*
-			 * bulk - source and bulk - drain diodes
-			 * here we just evaluate the ideal diode current and the
-			 * corresponding derivative (conductance).
-			 */
-            if (vbs <= 0)
-            {
-                CondBs = sourceSatCur / vt;
-                BsCurrent = CondBs * vbs;
-                CondBs += _baseConfig.Gmin;
-            }
-            else
-            {
-                var evbs = Math.Exp(vbs / vt);
-                CondBs = sourceSatCur * evbs / vt + _baseConfig.Gmin;
-                BsCurrent = sourceSatCur * (evbs - 1);
-            }
-            if (vbd <= 0)
-            {
-                CondBd = drainSatCur / vt;
-                BdCurrent = CondBd * vbd;
-                CondBd += _baseConfig.Gmin;
-            }
-            else
-            {
-                var evbd = Math.Exp(vbd / vt);
-                CondBd = drainSatCur * evbd / vt + _baseConfig.Gmin;
-                BdCurrent = drainSatCur * (evbd - 1);
-            }
-
-            /* now to determine whether the user was able to correctly
-			 * identify the source and drain of his device
-			 */
-            if (vds >= 0)
-            {
-                // normal mode
-                Mode = 1;
-            }
-            else
-            {
-                // inverse mode
-                Mode = -1;
-            }
-
+            
             {
                 /*
 				* this routine evaluates the drain current, its derivatives and
@@ -267,9 +112,9 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level2
                 /* 'local' variables - these switch d & s around appropriately
 				 * so that we don't have to worry about vds < 0
 				 */
-                var lvbs = Mode  > 0 ? vbs : vbd;
-                var lvds = Mode * vds;
-                var lvgs = Mode > 0 ? vgs : vgd;
+                var lvbs = vbs;
+                var lvds = vds;
+                var lvgs = vgs;
                 var phiMinVbs = _temp.TempPhi - lvbs;
                 double tmp; /* a temporary variable, not used for more than */
                             /* about 10 lines at a time */
@@ -758,65 +603,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level2
             doneval:
             Von = _mbp.MosfetType * von;
             SaturationVoltageDs = _mbp.MosfetType * vdsat;
-
-            /*
-			 * COMPUTE EQUIVALENT DRAIN CURRENT SOURCE
-			 */
-            DrainCurrent = Mode * cdrain - BdCurrent;
-
-            /*
-			 * check convergence
-			 */
-            if (!_bp.Off || state.Init != InitializationModes.Fix)
-            {
-                if (check == 1)
-                    state.IsConvergent = false;
-            }
-
-            VoltageBs = vbs;
-            VoltageBd = vbd;
-            VoltageGs = vgs;
-            VoltageDs = vds;
-
-            /*
-			 * load current vector
-			 */
-            var ceqbs = _mbp.MosfetType * (BsCurrent - (CondBs - _baseConfig.Gmin) * vbs);
-            var ceqbd = _mbp.MosfetType * (BdCurrent - (CondBd - _baseConfig.Gmin) * vbd);
-            if (Mode >= 0)
-            {
-                xnrm = 1;
-                xrev = 0;
-                cdreq = _mbp.MosfetType * (cdrain - CondDs * vds - Transconductance * vgs - TransconductanceBs * vbs);
-            }
-            else
-            {
-                xnrm = 0;
-                xrev = 1;
-                cdreq = -_mbp.MosfetType * (cdrain - CondDs * -vds - Transconductance * vgd - TransconductanceBs * vbd);
-            }
-            BulkPtr.Value -= ceqbs + ceqbd;
-            DrainPrimePtr.Value += ceqbd - cdreq;
-            SourcePrimePtr.Value += cdreq + ceqbs;
-
-            // Load Y-matrix
-            DrainDrainPtr.Value += _temp.DrainConductance;
-            SourceSourcePtr.Value += _temp.SourceConductance;
-            BulkBulkPtr.Value += CondBd + CondBs;
-            DrainPrimeDrainPrimePtr.Value += _temp.DrainConductance + CondDs + CondBd + xrev * (Transconductance + TransconductanceBs);
-            SourcePrimeSourcePrimePtr.Value += _temp.SourceConductance + CondDs + CondBs + xnrm * (Transconductance + TransconductanceBs);
-            DrainDrainPrimePtr.Value += -_temp.DrainConductance;
-            SourceSourcePrimePtr.Value += -_temp.SourceConductance;
-            BulkDrainPrimePtr.Value -= CondBd;
-            BulkSourcePrimePtr.Value -= CondBs;
-            DrainPrimeDrainPtr.Value += -_temp.DrainConductance;
-            DrainPrimeGatePtr.Value += (xnrm - xrev) * Transconductance;
-            DrainPrimeBulkPtr.Value += -CondBd + (xnrm - xrev) * TransconductanceBs;
-            DrainPrimeSourcePrimePtr.Value += -CondDs - xnrm * (Transconductance + TransconductanceBs);
-            SourcePrimeGatePtr.Value += -(xnrm - xrev) * Transconductance;
-            SourcePrimeSourcePtr.Value += -_temp.SourceConductance;
-            SourcePrimeBulkPtr.Value += -CondBs - (xnrm - xrev) * TransconductanceBs;
-            SourcePrimeDrainPrimePtr.Value += -CondDs - xrev * (Transconductance + TransconductanceBs);
+            return cdrain;
         }
     }
 }
